+++ checkov/terraform/graph_builder/local_graph.py
@@ -5,7 +5,7 @@ import os
 from collections import defaultdict
 from functools import partial
 from pathlib import Path
-from typing import List, Optional, Union, Any, Dict, Set, Tuple
+from typing import List, Optional, Union, Any, Dict, Set, Tuple, overload, TYPE_CHECKING
 
 from typing_extensions import TypedDict
 
@@ -36,6 +36,8 @@ from checkov.terraform.graph_builder.utils import (
 from checkov.terraform.graph_builder.utils import is_local_path
 from checkov.terraform.graph_builder.variable_rendering.renderer import TerraformVariableRenderer
 
+if TYPE_CHECKING:
+    from checkov.common.typing import TFDefinitionKeyType
 
 MODULE_RESERVED_ATTRIBUTES = ("source", "version")
 CROSS_VARIABLE_EDGE_PREFIX = '[cross-variable] '
@@ -56,8 +58,8 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         self.relative_paths_cache: dict[tuple[str, str], str] = {}
         self.abspath_cache: Dict[str, str] = {}
         self.dirname_cache: Dict[str, str] = {}
-        self.vertices_by_module_dependency_by_name: Dict[Tuple[str, str], Dict[BlockType, Dict[str, List[int]]]] = defaultdict(partial(defaultdict, partial(defaultdict, list)))
-        self.vertices_by_module_dependency: Dict[Tuple[str, str], Dict[BlockType, List[int]]] = defaultdict(partial(defaultdict, list))
+        self.vertices_by_module_dependency_by_name: Dict[TFModule | None, Dict[str, Dict[str, List[int]]]] = defaultdict(partial(defaultdict, partial(defaultdict, list)))  # type:ignore[arg-type]
+        self.vertices_by_module_dependency: Dict[TFModule | None, Dict[str, List[int]]] = defaultdict(partial(defaultdict, list))  # type:ignore[arg-type]
         self.enable_foreach_handling = strtobool(os.getenv('CHECKOV_ENABLE_FOREACH_HANDLING', 'True'))
         self.enable_modules_foreach_handling = strtobool(os.getenv('CHECKOV_ENABLE_MODULES_FOREACH_HANDLING', 'True'))
         self.use_new_tf_parser = strtobool(os.getenv('CHECKOV_NEW_TF_PARSER', 'True'))
@@ -85,18 +87,24 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             logging.info(f"Rendering variables, graph has {len(self.vertices)} vertices and {len(self.edges)} edges")
             renderer = TerraformVariableRenderer(self)
             renderer.render_variables_from_local_graph()
-            self.update_vertices_breadcrumbs_and_module_connections()
-            self.update_nested_modules_address()
+            self.update_vertices_fields()
             if strtobool(os.getenv("CHECKOV_EXPERIMENTAL_CROSS_VARIABLE_EDGES", "True")):
                 # experimental flag on building cross variable edges for terraform graph
                 logging.info("Building cross variable edges")
                 edges_count = len(self.edges)
                 self._build_cross_variable_edges()
                 logging.info(f"Found {len(self.edges) - edges_count} cross variable edges")
+        else:
+            self.update_vertices_fields()
+
+    def update_vertices_fields(self) -> None:
+        # Important to keep those 2 functions together, as the first affects the calculation of the second
+        self._update_vertices_breadcrumbs_and_module_connections()
+        self._update_nested_modules_address()
 
     def _create_vertices(self) -> None:
         logging.info("Creating vertices")
-        self.vertices = [None] * len(self.module.blocks)
+        self.vertices = [None] * len(self.module.blocks)  # type:ignore[list-item]  # are correctly set in the next lines
         for i, block in enumerate(self.module.blocks):
             self.vertices[i] = block
             self._add_block_data_to_graph(i, block)
@@ -117,8 +125,9 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             self.vertices_by_module_dependency[block.source_module_object][block.block_type].append(idx)
             self.vertices_by_module_dependency_by_name[block.source_module_object][block.block_type][block.name].append(idx)
         else:
-            self.vertices_by_module_dependency[(block.module_dependency, block.module_dependency_num)][block.block_type].append(idx)
-            self.vertices_by_module_dependency_by_name[(block.module_dependency, block.module_dependency_num)][block.block_type][block.name].append(idx)
+            # mypy: fixed, when the old parser is removed
+            self.vertices_by_module_dependency[(block.module_dependency, block.module_dependency_num)][block.block_type].append(idx)  # type:ignore[index]
+            self.vertices_by_module_dependency_by_name[(block.module_dependency, block.module_dependency_num)][block.block_type][block.name].append(idx)  # type:ignore[index]
 
         self.in_edges[idx] = []
         self.out_edges[idx] = []
@@ -126,10 +135,10 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
     def _arrange_graph_data(self) -> None:
         # reset all the relevant data
         self.vertices_by_block_type = defaultdict(list)
-        self.vertices_block_name_map = defaultdict(partial(defaultdict, list))
+        self.vertices_block_name_map = defaultdict(partial(defaultdict, list))  # type:ignore[arg-type]
         self.map_path_to_module = {}
-        self.vertices_by_module_dependency = defaultdict(partial(defaultdict, list))
-        self.vertices_by_module_dependency_by_name = defaultdict(partial(defaultdict, partial(defaultdict, list)))
+        self.vertices_by_module_dependency = defaultdict(partial(defaultdict, list))  # type:ignore[arg-type]
+        self.vertices_by_module_dependency_by_name = defaultdict(partial(defaultdict, partial(defaultdict, list)))  # type:ignore[arg-type]
         self.edges = []
         for i in range(len(self.vertices)):
             self.out_edges[i] = []
@@ -138,7 +147,7 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         for i, block in enumerate(self.vertices):
             self._add_block_data_to_graph(i, block)
 
-    def _get_aliases(self) -> Dict[str, Dict[str, BlockType]]:
+    def _get_aliases(self) -> Dict[str, Dict[str, str]]:
         """
         :return aliases: map between alias names that are found inside the blocks and the block type their aliased to.
         """
@@ -170,7 +179,11 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
                     break
             return
 
-        block_dirs_to_modules: Dict[Tuple[str, str], Dict[str, Set[int]]] = defaultdict(dict)
+        if not self.module.module_dependency_map:
+            # no need to proceed further
+            return
+
+        block_dirs_to_modules: Dict[Tuple[str, str | None], Dict[str, Set[int]]] = defaultdict(dict)
         for dir_name, paths_to_modules in self.module.module_dependency_map.items():
             # for each directory, find the module vertex that imported it
             for path_to_module in paths_to_modules:
@@ -199,9 +212,9 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
 
         for vertex in self.vertices:
             # match the right module vertex according to the vertex path directory
-            module_dependency_nums = block_dirs_to_modules.get((self.get_dirname(vertex.path), vertex.module_dependency))
+            module_dependency_nums = block_dirs_to_modules.get((self.get_dirname(vertex.path), vertex.module_dependency))  # type:ignore[arg-type]  # will be fixed when removing terraform/checks from mypy exclusion
             if module_dependency_nums:
-                module_indices = module_dependency_nums.get(vertex.module_dependency_num)
+                module_indices = module_dependency_nums.get(vertex.module_dependency_num)  # type:ignore[arg-type]  # vertex.module_dependency_num can be None, which is ok
                 if module_indices:
                     vertex.source_module = module_indices
 
@@ -213,7 +226,7 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         for origin_node_index, vertex in enumerate(self.vertices):
             self._build_edges_for_vertex(origin_node_index, vertex, aliases, resources_types)
 
-    def _build_edges_for_vertex(self, origin_node_index: int, vertex: TerraformBlock, aliases: Dict[str, Dict[str, BlockType]],
+    def _build_edges_for_vertex(self, origin_node_index: int, vertex: TerraformBlock, aliases: Dict[str, Dict[str, str]],
                                 resources_types: List[str], cross_variable_edges: bool = False,
                                 referenced_modules: Optional[List[Dict[str, Any]]] = None) -> None:
 
@@ -244,12 +257,15 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
                             referenced_module_object = module.get("source_module_object")
                             if not self.use_new_tf_parser:
                                 source_module_object = referenced_module_object if source_module_object else None
-                            dest_node_index = self._find_vertex_index_relative_to_path(
-                                vertex_reference.block_type, reference_name, referenced_module_path,
-                                vertex.module_dependency,
-                                vertex.module_dependency_num, referenced_module_idx,
-                                source_module_object=source_module_object
-                            )
+                            if referenced_module_path is None:
+                                dest_node_index = -1
+                            else:
+                                dest_node_index = self._find_vertex_index_relative_to_path(
+                                    vertex_reference.block_type, reference_name, referenced_module_path,
+                                    vertex.module_dependency,
+                                    vertex.module_dependency_num, referenced_module_idx,
+                                    source_module_object=source_module_object
+                                )
                             self._create_edge_from_reference(attribute_key, origin_node_index, dest_node_index,
                                                              sub_values, vertex_reference, cross_variable_edges)
                     if vertex.module_dependency or hasattr(vertex, "source_module_object"):
@@ -324,11 +340,11 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
                 if self.get_dirname(self.vertices[index].path) == dest_module_path
             ]
         else:
-            target_path = get_path_with_nested_modules(vertex)
+            target_path_old = get_path_with_nested_modules(vertex)
             return [
                 index
-                for index in self.vertices_by_module_dependency.get(
-                    (target_path, self.module.module_address_map.get((vertex.path, vertex.name))), {}).get(
+                for index in self.vertices_by_module_dependency.get(  # type:ignore[call-overload]  # fixed, when the old parser is removed
+                    (target_path_old, self.module.module_address_map.get((vertex.path, vertex.name))), {}).get(
                     BlockType.VARIABLE, [])
                 if self.get_dirname(self.vertices[index].path) == dest_module_path
             ]
@@ -416,9 +432,16 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         return ""
 
     def _find_vertex_index_relative_to_path(
-        self, block_type: BlockType, name: str, block_path: str, module_path: str, module_num: str, relative_module_idx: Optional[int] = None, source_module_object: Optional[TFModule] = None
+        self,
+        block_type: str,
+        name: str,
+        block_path: str,
+        module_path: TFDefinitionKeyType | None,
+        module_num: str | None,
+        relative_module_idx: Optional[int] = None,
+        source_module_object: Optional[TFModule] = None,
     ) -> int:
-        relative_vertices = []
+        relative_vertices: list[int] = []
         if self.use_new_tf_parser:
             if relative_module_idx is None:
                 module_dependency_by_name_key = source_module_object
@@ -431,8 +454,10 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             module_dependency_by_name_key = next(k for k, v in self.vertices_by_module_dependency.items() if
                                                  v.get(BlockType.MODULE, []).__contains__(relative_module_idx))
         else:
-            module_dependency_by_name_key = (module_path, module_num)
-        possible_vertices = self.vertices_by_module_dependency_by_name.get(module_dependency_by_name_key, {}).get(block_type, {}).get(name, [])  # important to use this specific map for big graph performance
+            module_dependency_by_name_key = (module_path, module_num)  # type:ignore[assignment]  # fixed, when the old parser is removed
+
+        # important to use this specific map for big graph performance
+        possible_vertices = self.vertices_by_module_dependency_by_name.get(module_dependency_by_name_key, {}).get(block_type, {}).get(name, [])
         for vertex_index in possible_vertices:
             vertex = self.vertices[vertex_index]
             if self.get_dirname(vertex.path) == self.get_dirname(block_path):
@@ -478,9 +503,14 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         vertex_index: int,
         attribute_key: str,
         attribute_value: Any,
-        change_origin_id: int,
+        change_origin_id: int | None,
         attribute_at_dest: Optional[Union[str, List[str]]],
+        transform_step: bool = False,
     ) -> None:
+        if change_origin_id is None:
+            # no need to proceed further
+            return
+
         previous_breadcrumbs = []
         attribute_at_dest = self.vertices[change_origin_id].find_attribute(attribute_at_dest)
         if attribute_at_dest:
@@ -495,7 +525,8 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             changed_attributes = filter_sub_keys(changed_attributes)
             self.update_vertex_config(vertex, changed_attributes)
 
-    def update_vertex_config(self, vertex: TerraformBlock, changed_attributes: Union[List[str], Dict[str, Any]], dynamic_blocks: bool = False) -> None:
+    @staticmethod
+    def update_vertex_config(vertex: TerraformBlock, changed_attributes: Union[List[str], Dict[str, Any]], dynamic_blocks: bool = False) -> None:
         if not changed_attributes:
             # skip, if there is no change
             return
@@ -525,13 +556,16 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
 
         if len(changed_attributes) > 0:
             if vertex.block_type == BlockType.LOCALS:
-                updated_config = updated_config.get(vertex_name)
+                updated_local_config = updated_config.get(vertex_name)
+                update_dictionary_attribute(vertex.config, vertex_name, updated_local_config, dynamic_blocks)
+                return
+
             update_dictionary_attribute(vertex.config, vertex_name, updated_config, dynamic_blocks)
 
     def get_resources_types_in_graph(self) -> List[str]:
         return self.module.get_resources_types()
 
-    def update_vertices_breadcrumbs_and_module_connections(self) -> None:
+    def _update_vertices_breadcrumbs_and_module_connections(self) -> None:
         """
         The function processes each vertex's breadcrumbs:
         1. Get more data to each vertex in breadcrumb (name, path, hash and type)
@@ -589,8 +623,11 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             self.abspath_cache[path] = dir_name
         return dir_name
 
-    def update_nested_modules_address(self) -> None:
+    def _update_nested_modules_address(self) -> None:
         for vertex in self.vertices:
+            if vertex.attributes.get(CustomAttributes.TF_RESOURCE_ADDRESS) is not None:
+                # Can happen for example in `tf_plan` files as the address already exists
+                continue
             if vertex.block_type not in parser_registry.context_parsers:
                 continue
             source_module = vertex.breadcrumbs.get(CustomAttributes.SOURCE_MODULE)
@@ -614,6 +651,9 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
 
     def _should_add_edge(self, vertex: TerraformBlock, dest_module_path: str, module_node: TerraformBlock) -> bool:
         if self.use_new_tf_parser:
+            if not vertex.source_module_object:
+                return False
+
             return (self.get_dirname(vertex.path) == dest_module_path) and \
                 (
                     vertex.source_module_object == module_node.source_module_object  # The vertex is in the same file
@@ -621,11 +661,12 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
                     == self.get_abspath(module_node.path)  # The vertex is in the correct dependency path)
             )
         else:
-            return (self.get_dirname(vertex.path) == dest_module_path) and \
-                (
-                    vertex.module_dependency == module_node.module_dependency  # The vertex is in the same file
-                    or self.get_abspath(vertex.module_dependency)
-                    == self.get_abspath(module_node.path)  # The vertex is in the correct dependency path
+            return (self.get_dirname(vertex.path) == dest_module_path) and (
+                vertex.module_dependency == module_node.module_dependency  # The vertex is in the same file
+                or (
+                    vertex.module_dependency is not None
+                    and self.get_abspath(vertex.module_dependency) == self.get_abspath(module_node.path)  # type:ignore[arg-type]  # old flow, will be removed
+                )  # The vertex is in the correct dependency path
             )
 
 
@@ -638,6 +679,20 @@ def to_list(data: Any) -> list[Any] | dict[str, Any]:
         return {key: to_list(val) for key, val in data.items()}
     else:
         return [data]
+
+
+@overload
+def update_dictionary_attribute(
+        config: dict[str, Any], key_to_update: str, new_value: Any, dynamic_blocks: bool = False
+) -> dict[str, Any]:
+    ...
+
+
+@overload
+def update_dictionary_attribute(
+        config: list[Any], key_to_update: str, new_value: Any, dynamic_blocks: bool = False
+) -> list[Any]:
+    ...
 
 
 def update_dictionary_attribute(
@@ -705,8 +760,8 @@ def get_path_with_nested_modules(block: TerraformBlock) -> str:
     if not block.module_dependency:
         return block.path
     if not strtobool(os.getenv('CHECKOV_ENABLE_NESTED_MODULES', 'True')):
-        return unify_dependency_path([block.module_dependency, block.path])
-    return get_tf_definition_key_from_module_dependency(block.path, block.module_dependency, block.module_dependency_num)
+        return unify_dependency_path([block.module_dependency, block.path])  # type:ignore[no-any-return]  # will be fixed when removing terraform/checks from mypy exclusion
+    return get_tf_definition_key_from_module_dependency(block.path, block.module_dependency, block.module_dependency_num)  # type:ignore[arg-type]  # will be fixed when removing terraform/checks from mypy exclusion
 
 
 def get_vertex_as_tf_module(block: TerraformBlock) -> TFModule: